inject file assembly directives
Currently, non-IA builds are not reproducibile since build paths are
being injected into the debug symbols. These are coming from the use of
.S assembler files during the glibc build. No STT_FILE section is added
during the assembly but when linking, ld decides to add one to aid
debugging and ensure references between the different object files its
linking remain clear.

We can avoid this by injecting a file header into the assembler files
ahead of time, choosing a filename which does not contain build system
paths.

This is a bit of a workaround/hack but does significantly reduce the
build system references in target binaries for the non-IA architectures
which use .S files.

RP
2018/10/3

Upstream-Status: Pending

Signed-off-by: Khem Raj <raj.khem@gmail.com>